Forgot to push again
On program start, LevelDataSOHandler loads up LevelSOList.json to find and confirm the corresponding LevelScriptableObject on Resources. if not found, it loads up the master LevelData.Json and finds the LevelID, and creates a LevelScriptableObject asset with its appropriate data.

TODO: same functionality (or a generic class/method) for Patterns and Projectiles.

~rearranged some files again
+Levels.json
+LevelSOList.json
+LevelData.cs
+LevelSO.cs
+LevelDataSOHandler.cs
~ProjectileJSONLoader to generic JsonLoader
~buncha additions and edits in FileNames.cs
~+-prolly a buncha other stuff i forgot